refactored split modules outside of __init__
