+++ .travis.yml
@@ -2,8 +2,8 @@ language: python
 python:
 - '2.7'
 - '3.5'
-install: pip install .
-script: python -m tests.test_phabricator
+install: pip install . ; pip install pytest
+script: pytest
 deploy:
   provider: pypi
   user: disqus
