+++ requirements.txt
@@ -1,3 +1,3 @@
-setuptools>=33.1.1
-unittest2>=1.1.0
-mock>=2.0.0
+setuptools==41.0.1
+mock==3.0.5
+unittest2==1.1.0
